Safari-609.1.11.1/LayoutTests/http/tests/xmlhttprequest/basic-auth-credentials-escaping-expected.txt

Tests for XMLHttpRequest authentication.

async: User: 1, password: %a%.
async2: User: 2, password: @.
async3: User: 3, password: @.
async4: User: 4, password: :.
async5: User: 5, password: .
async6: User: 6, password: a.
async7: User: 7, password: /@\ "#><?`.
